refactoring punctuation, vocab into modules
