+++ wordsiv/__init__.py
@@ -1,15 +1,16 @@
 """WordSiv is a Python library for generating proofing text with a limited character set."""
 
-from functools import cached_property, lru_cache
+from functools import lru_cache
 from itertools import accumulate
 import logging
 import random
-import regex
 import json
-from math import exp
-from . import vocab_data
 from importlib import resources
-from .filter import FilterError, BIG_NUM, filter_wordcount
+
+from .vocab import Vocab
+from .filter import FilterError, BIG_NUM
+from .punctuation import DEFAULT_PUNCTUATION, punctuate
+from . import vocab_data
 
 log = logging.getLogger(__name__)
 
@@ -26,136 +27,10 @@ DEFAULT_VOCABS = {
     "fa": ("fa_subs_meta.json", "fa_subs.tsv"),
 }
 
-DEFAULT_PUNCTUATION = {
-    "en": {
-        "start": {"": 1.0},
-        "insert": {
-            " ": 0.365,
-            ", ": 0.403,
-            ": ": 0.088,
-            "; ": 0.058,
-            "–": 0.057,
-            "—": 0.022,
-            " … ": 0.006,
-        },
-        "wrap_sent": {
-            ("", "."): 0.923,
-            ("", "!"): 0.034,
-            ("", "?"): 0.04,
-            ("", "…"): 0.003,
-        },
-        "wrap_inner": {
-            ("", ""): 0.825,
-            ("(", ")"): 0.133,
-            ("‘", "’"): 0.013,
-            ("“", "”"): 0.028,
-        },
-    },
-    "ar": {
-        "insert": {" ": 0.364, ": ": 0.108, "، ": 0.463, "؛ ": 0.066},
-        "wrap_sent": {("", "."): 0.914, ("", "؟"): 0.052, ("", "!"): 0.033},
-        "wrap_inner": {("", ""): 0.971, ("’", "‘"): 0.007, ("”", "“"): 0.022},
-    },
-    "fa": {
-        "insert": {" ": 0.364, ": ": 0.108, "، ": 0.463, "؛ ": 0.066},
-        "wrap_sent": {("", "."): 0.914, ("", "؟"): 0.052, ("", "!"): 0.033},
-        "wrap_inner": {("", ""): 0.971, ("’", "‘"): 0.007, ("”", "“"): 0.022},
-    },
-    "es": {
-        "start": {"": 100},
-        "insert": {
-            " ": 0.277,
-            ", ": 0.49,
-            ": ": 0.093,
-            "; ": 0.073,
-            "– ": 0.026,
-            "— ": 0.03,
-            "… ": 0.011,
-        },
-        "wrap_sent": {
-            ("", "."): 0.928,
-            ("¡", "!"): 0.029,
-            ("¿", "?"): 0.036,
-            ("", "…"): 0.008,
-        },
-        "wrap_inner": {
-            ("", ""): 0.814,
-            ("(", ")"): 0.129,
-            ("‘", "’"): 0.013,
-            ("“", "”"): 0.044,
-        },
-    },
-}
-
-
-def _punctuate(
-    punctuation: dict,
-    rand: random.Random,
-    words: list[str],
-    glyphs: str,
-    rnd_punc: float,
-):
-    """Punctuate a list of words and join into a sentence using a punctuation dict"""
-
-    def random_available(option_weight, glyphs: str, rand, rnd_punc: float):
-        punc_prob = {
-            (punc, (1 - rnd_punc) * prob + rnd_punc * 1)
-            for punc, prob in option_weight.items()
-            # if glyphs is set, check if we have the glyphs we need to punctuate
-            # we aren't strict about having spaces, hence glyphs + ' '
-            if not glyphs or all(c in glyphs + " " for c in punc)
-        }
-
-        if punc_prob:
-            options, weights = zip(*punc_prob)
-            return rand.choices(options, weights=weights, k=1)[0]
-        else:
-            return None
-
-    insert = random_available(punctuation["insert"], glyphs, rand, rnd_punc) or ""
-    wrap_sent = random_available(punctuation["wrap_sent"], glyphs, rand, rnd_punc) or (
-        "",
-        "",
-    )
-    wrap_inner = random_available(
-        punctuation["wrap_inner"], glyphs, rand, rnd_punc
-    ) or ("", "")
-
-    if len(words) > 2:
-        tokens = []
-
-        # place inner wrap punctuation on words (parentheses, quotes, etc)
-        wrap_left_index = rand.randrange(0, len(words) - 1)
-        wrap_right_index = rand.randrange(wrap_left_index, len(words) - 1)
-        words[wrap_left_index] = wrap_inner[0] + words[wrap_left_index]
-        words[wrap_right_index] = words[wrap_right_index] + wrap_inner[1]
-
-        # add insert punctuation and spaces
-        # pick an index that hasn't already been used (simpler)
-        insert_index = rand.choice(
-            [
-                n
-                for n in range(1, len(words) - 1)
-                if n not in [words[wrap_left_index], words[wrap_right_index]]
-            ]
-        )
-
-        # put "insert" punctuation in place of a space
-        separators = [" "] * (len(words) - 1)
-        separators[insert_index - 1] = insert
-        sent = "".join(
-            [t for pair in zip(words, separators) for t in pair] + [words[-1]]
-        )
-    else:
-        sent = " ".join(words)
-
-    # place surrounding punctuation
-    return wrap_sent[0] + sent + wrap_sent[1]
-
 
 @lru_cache(maxsize=None)
 def _accumulate_weights(counts: list[int]) -> list[int]:
-    """Accumulate weights and cache results for speed"""
+    """Accumulate weights and cache results"""
 
     return list(accumulate(counts))
 
@@ -164,7 +39,7 @@ def _accumulate_weights(counts: list[int]) -> list[int]:
 def _split_wordcount(
     word_count: list[tuple[str, int]]
 ) -> tuple[tuple[str], tuple[int]]:
-    """Split a wordcount list into words and counts."""
+    """Split a wordcount list into words and counts and cache results"""
 
     return tuple(i[0] for i in word_count), tuple(i[1] for i in word_count)
 
@@ -188,92 +63,6 @@ def _sample_word(
     accumulated_counts = _accumulate_weights(adjusted_counts)
 
     return rand.choices(words, cum_weights=accumulated_counts)[0]
-
-
-class VocabEmptyError(Exception):
-    pass
-
-
-class VocabFormatError(Exception):
-    pass
-
-
-class Vocab:
-    """A vocabulary of words and occurrence counts with metadata for filtering and punctuating.
-
-    Attributes:
-        bicameral (bool): Specifies whether the vocabulary has uppercase and lowercase letters.
-        data (str): A TSV-formatted string with word, count pairs or a newline-delimited list of words.
-        data_file (str): A path to a file to supply the data instead of the data attribute
-        punctuation (Punctuation): A Punctuation object for handling punctuation in generated text.
-    """
-
-    def __init__(
-        self,
-        lang: str,
-        bicameral: bool,
-        punctuation: dict = None,
-        data: str = None,
-        data_file: str = None,
-        meta: dict = None,
-    ):
-        """Initializes the Vocab instance."""
-
-        self.lang = lang
-        self.bicameral = bicameral
-        self.punctuation = punctuation
-        self._data = data
-        self.data_file = data_file
-        self.meta = meta
-
-        if data and data_file:
-            raise ValueError("Cannot specify both 'data' and 'data_file'")
-        elif data is None and not data_file:
-            raise ValueError("Must specify either 'data' or 'data_file'")
-
-    @cached_property
-    def data(self):
-        """Returns raw data from parameter _data or data_file."""
-
-        if self._data is not None:
-            data = self._data
-        elif getattr(self, "data_file", None):
-            with open(self.data_file, "r", encoding="utf8") as f:
-                data = f.read()
-        if not data:
-            raise VocabEmptyError(f"No data found in {self.data_file}")
-
-        return data
-
-    @cached_property
-    def wordcount_str(self) -> str:
-        """Returns a TSV-formatted string with words and counts."""
-
-        firstline = self.data.partition("\n")[0]
-
-        if regex.match(r"[[:alpha:]]+\t\d+$", firstline):
-            # if we have counts, return the original string
-            return self.data
-        elif regex.match(r"[[:alpha:]]+$", firstline):
-            # if we just have newline-delimited words, add counts of 1
-            return "\n".join(f"{w}\t1" for w in self.data.splitlines())
-        else:
-            raise VocabFormatError(
-                "The vocab file is formatted incorrectly. "
-                "Should be a TSV file with words and counts as columns, or a newline-delimited list of words."
-            )
-
-    @cached_property
-    def wordcount(self) -> list[tuple[str, int]]:
-        """Returns a list of tuples with words and counts."""
-
-        return [
-            (l.split()[0], int(l.split()[1])) for l in self.wordcount_str.splitlines()
-        ]
-
-    @lru_cache(maxsize=None)
-    def filter(self, *args, **kwargs):
-        return filter_wordcount(self.wordcount_str, self.bicameral, *args, **kwargs)
 
 
 class WordSiv:
@@ -616,7 +405,7 @@ class WordSiv:
                 except KeyError:
                     return " ".join(words)
 
-            return _punctuate(
+            return punctuate(
                 punctuation,
                 self.rand,
                 word_list,
